// web/src/api/generated/models/workflow-trigger-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NovaCloudEdu API
 * NovaCloudEdu Smart Education Platform Backend API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 工作流触发器响应
 */
export interface WorkflowTriggerResponse {
    /**
     * 触发器ID
     */
    'id'?: number;
    /**
     * 工作流ID
     */
    'workflowId'?: number;
    /**
     * 触发器类型：SCHEDULE/WEBHOOK/EVENT
     */
    'type'?: string;
    /**
     * 触发器名称
     */
    'name'?: string;
    /**
     * 是否启用
     */
    'enabled'?: boolean;
    /**
     * 配置JSON
     */
    'config'?: { [key: string]: object; };
    /**
     * 最后触发时间
     */
    'lastTriggeredAt'?: string;
    /**
     * 触发次数
     */
    'triggerCount'?: number;
    /**
     * 创建时间
     */
    'createTime'?: string;
}